Move rendering methods from the order class

+++ BikeDistributor/Bike.cs
@@ -2,11 +2,11 @@
 {
     public class Bike
     {
-        public const int OneThousand = 1000;
-        public const int TwoThousand = 2000;
-        public const int FiveThousand = 5000;
+        public const decimal OneThousand = 1000;
+        public const decimal TwoThousand = 2000;
+        public const decimal FiveThousand = 5000;
     
-        public Bike(string brand, string model, int price)
+        public Bike(string brand, string model, decimal price)
         {
             Brand = brand;
             Model = model;
@@ -15,6 +15,6 @@
 
         public string Brand { get; private set; }
         public string Model { get; private set; }
-        public int Price { get; set; }
+        public decimal Price { get; set; }
     }
 }
